Naviagation and Dialog refactioring
Refactoring navigation to use layout files instead of creating the views each time.

Moving disclaimer dialog to be handled in its own java class
