+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@color/colorPrimaryDark"
     android:clickable="true"
     android:focusable="true"
     tools:context=".MainActivity">
@@ -53,14 +52,13 @@
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
                 android:layout_weight="1"
-                android:background="@color/colorPrimaryDark"
                 android:clickable="true"
+                android:focusable="true"
                 android:onClick="changePage"
                 android:tag="@string/advice"
                 android:text="@string/advice"
                 android:textSize="18dp"
-                android:textColor="@color/colorAccent"
-                android:focusable="true"></Button>
+                android:textColor="@color/colorAccent"></Button>
         </LinearLayout>
 
         <LinearLayout
@@ -75,36 +73,9 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent">
 
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:orientation="vertical">
-
-                <ImageView
-                    android:id="@+id/imageView"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_gravity="center_horizontal"
-                    android:clickable="false"
-                    android:paddingTop="50dp"
-                    app:layout_constraintEnd_toEndOf="parent"
-                    app:layout_constraintStart_toEndOf="@+id/contactLayout"
-                    app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toTopOf="parent"
-                    app:srcCompat="@mipmap/gibsams_logo" />
-
-                <TextView
-                    android:id="@+id/textView3"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:textAlignment="center"
-                    android:paddingTop="25dp"
-                    android:paddingBottom="20dp"
-                    android:textColor="@color/colorPrimaryDark"
-                    android:textSize="32sp"
-                    android:text="@string/slogan"
-                    android:textStyle="italic"/>
-            </LinearLayout>
+            <include
+                android:id="@+id/layoutView"
+                layout="@layout/home_page"/>
 
         </LinearLayout>
     </LinearLayout>
